moved MeliDialog
