+++ exampleApp/src/main/java/com/mercadolibre/android/ui/example/ui/widgets/dialog/DummyFeedbackDialog.java
@@ -6,7 +6,7 @@ import android.widget.Toast;
 
 import com.mercadolibre.android.ui.example.R;
 import com.mercadolibre.android.ui.widgets.ErrorView;
-import com.mercadolibre.android.ui.widgets.MeliDialog;
+import com.mercadolibre.android.ui.widgets.dialogs.MeliDialog;
 
 public class DummyFeedbackDialog extends MeliDialog {
